fix: 🐛 test for next/image test (#1552)
this is jest warning:
Warning: React does not recognize the `fetchPriority` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `fetchpriority` instead. If you accidentally passed it from a parent component, remove it from the DOM element.

see here: https://github.com/testing-library/react-testing-library/issues/1324